split things into files, move into src/lib
