+++ src/app/page.tsx
@@ -1,6 +1,6 @@
-import Header from "./Header";
-import About from "./About";
-import App from "./App"
+import Header from "src/lib/Header";
+import About from "src/lib/About";
+import App from "src/lib/App"
 
 export default function Home() {
   return (
